Closes #2
Basic ingest program is complete, it isn't integrated as a full module
which I need to get to and the code quality isn't excellent but it'll
get the job done. it loops through the directories and subdirectories
of all the stocknet data, splits the individual files into json objects,
decodes the json objects using a custom function to class Tweet, and
then pushes the class's dictionaries en masse to mongodb.